+++ include/Actions/PowerAvg.h
@@ -8,12 +8,14 @@ namespace Actions {
 
 class PowerAvg : public NumericAction {
    public:
-    PowerAvg(const char* str, int param);
+    PowerAvg(const char* str, size_t param);
     void exec(Print* p);
 };
 
-PowerAvg::PowerAvg(const char* str, int param) : NumericAction(str, param) {}
+PowerAvg::PowerAvg(const char* str, size_t param) : NumericAction(str, param) {}
 
-void PowerAvg::exec(Print* p) { ina231_set_avg((uint8_t)param); }
+void PowerAvg::exec(Print* p) { 
+    ina231_set_avg((uint8_t)param); 
+}
 
 }  // namespace Actions
